Extract Configuration into separate file
This change requires a node version > 6.0. This should be trivial to
update, and I see no reason why we'd want to stick to a lower version.

It's much easier to see what's going on when we do this. You can see the
defaults are all being set in the configuration, and we only ever need
the one function that the CLI cares about. All the settings are
contained on an object that is passed around, rather than being set on
every command runtime.